+++ src/main/java/edu/ie3/simona/api/data/em/ExtEmDataConnection.java
@@ -7,16 +7,18 @@
 package edu.ie3.simona.api.data.em;
 
 import edu.ie3.datamodel.models.value.PValue;
-import edu.ie3.simona.api.data.ExtData;
-import edu.ie3.simona.api.data.ExtInputDataContainer;
+import edu.ie3.datamodel.models.value.Value;
+import edu.ie3.simona.api.data.ExtInputDataConnection;
 import edu.ie3.simona.api.data.em.ontology.EmDataMessageFromExt;
 import edu.ie3.simona.api.data.em.ontology.ProvideEmSetPointData;
 import edu.ie3.simona.api.data.ontology.ScheduleDataServiceMessage;
 import java.util.*;
+import java.util.stream.Collectors;
 import org.apache.pekko.actor.ActorRef;
+import org.slf4j.Logger;
 
 /** Enables data connection of em data between SIMONA and SimonaAPI */
-public class ExtEmData implements ExtData {
+public class ExtEmDataConnection implements ExtInputDataConnection {
 
   /** Actor reference to service that handles ev data within SIMONA */
   private ActorRef emDataService;
@@ -27,14 +29,31 @@ public class ExtEmData implements ExtData {
   /** Assets that provide primary data to SIMONA */
   private final Map<String, UUID> extEmMapping;
 
-  public ExtEmData(Map<String, UUID> extEmMapping) {
+  public ExtEmDataConnection(Map<String, UUID> extEmMapping) {
     this.extEmMapping = extEmMapping;
   }
 
-  /** Sets the actor refs for data and control flow */
+  @Override
   public void setActorRefs(ActorRef emDataService, ActorRef extSimAdapter) {
     this.emDataService = emDataService;
     this.extSimAdapter = extSimAdapter;
+  }
+
+  public void convertAndSend(
+      long tick, Map<String, Value> data, Optional<Long> maybeNextTick, Logger log) {
+    // filtering the data and converting the keys
+    Map<UUID, PValue> convertedMap =
+        data.entrySet().stream()
+            .filter(e -> extEmMapping.containsKey(e.getKey()))
+            .collect(
+                Collectors.toMap(e -> extEmMapping.get(e.getKey()), e -> (PValue) e.getValue()));
+
+    if (convertedMap.isEmpty()) {
+      log.warn("No em data found! Sending no em data to SIMONA for tick {}.", tick);
+    } else {
+      log.debug("Provided SIMONA with em data.");
+      provideEmData(tick, convertedMap, maybeNextTick);
+    }
   }
 
   /** Returns a list of the uuids of the em agents that expect external set points */
@@ -58,27 +77,5 @@ public class ExtEmData implements ExtData {
     emDataService.tell(msg, ActorRef.noSender());
     // we need to schedule data receiver activation with scheduler
     extSimAdapter.tell(new ScheduleDataServiceMessage(emDataService), ActorRef.noSender());
-  }
-
-  /** Converts an input data package from an external simulation to a map of set points */
-  public Map<UUID, PValue> convertExternalInputToEmSetPoints(
-      ExtInputDataContainer extInputDataContainer) {
-    Map<UUID, PValue> emDataForSimona = new HashMap<>();
-    extInputDataContainer
-        .getSimonaInputMap()
-        .forEach(
-            (id, value) -> {
-              if (extEmMapping.containsKey(id)) {
-                if (value instanceof PValue pValue) {
-                  emDataForSimona.put(extEmMapping.get(id), pValue);
-                } else {
-                  throw new IllegalArgumentException("EmData can only handle PValue's!");
-                }
-              } else {
-                throw new IllegalArgumentException(
-                    "Input for asset with id " + id + " was provided, but it wasn't requested!");
-              }
-            });
-    return emDataForSimona;
   }
 }